npc.php: remove call to do_voodoo
We are basically not using _any_ of the preprocessing in `do_voodoo`,
because it mostly has to do with AJAX, sector locks, and redirects.
We don't need/want any of that in the NPC script, and so we can just
process the current Page directly.

Since "processing" pages never return, but "display" pages do, we
need to throw an exception after processing the page to ensure that
we're not accidentally processing a display page.