SmrPlayer: change return type of getShip method
We want to be able to call `update()` on the object that is returned
by `SmrPlayer::getShip()`, but the base class `AbstractSmrShip` does
not have this method.

We could have made the return type `SmrShip` instead, but this would
not work with the `DummyPlayer::getShip` method since `DummyShip` is
not (and cannot be) a child class of `SmrShip`.

Therefore, we keep the low-level return type in `SmrPlayer::getShip`
and simply add a method `AbstractSmrShip::update` that throws an
Exception if it is called. This can't just be an abstract method
because we actually instantiate the `AbstractSmrShip` class (though
primarily for testing). Fundamentally, `SmrShip` is more of a data
supplier class for the `AbstractSmrShip` class, which is the way we
should be thinking about this when we eventually restructure all of
the "Abstract" classes.

This fixes the phpstan warnings:

> Call to an undefined method AbstractSmrShip::update().
> Call to an undefined method AbstractSmrShip::cacheDummyShip().