Use PHPStan conditional return types in docstrings
> Comparison operation ">" between int<1, max> and array|int results in an error.

There are several methods that return an int when its argument is
specified, but an array when its argument is omitted. While these are
conveniently compact methods, the php-native union type `array|int` is
not narrow enough to encapsulate how the return type interacts with
the argument.

PHPStan provides a "conditional return type" syntax, which allows us
to codify this relationship. It is an advanced return type, specific
to PHPStan.

We also need to declare the type of a variable extracted from `$var`,
since there is no way for PHPStan to infer its type otherwise.
(Perhaps we need `Page::getInt` and friends?)