ChessGame: fix phpstan warning in tryAlgebraicMove
> Binary operation "-" between 8 and string results in an error.

It is not an error to use binary operators between numeric strings
and numbers, but PHPStan had no way of knowing that the string `$move`
would be of the form, e.g. "b2b4", and thus we can declare that the
2nd and 4th characters will be of the type `numeric-string`.

This also seems to fix an error with converting the input ranks to
y-coordinates. Using `8 - $rank` gives an incorrect mirrored value.

Also simplify the sanity check that the algebraic move is valid by
performing the check _after_ conversion to xy-coordinates.