phpcs.xml: ignore sniff DocBlockTagTypes.Unknown
Lots of tools use their own docstring tags, and PSR2R is not updated
frequently enough to catch all the new tags that are added. Some of
this has been fixed in php-fig-rectified/psr2r-sniffer#28, but no new
release has been made.

It's not a terribly high-value sniff, so just suppress it.