Fix phpstan warnings: unneeded elseif
> Elseif condition is always true.

PHPStan can detect when an else-if condition is always true, and
suggests changing it to an else branch in this case.

While this is technically more correct, it can sometimes reduce the
readability of the code. Worse, it can make the code susceptible to
doing the wrong thing if the surrounding code is modified in the
future. For this reason, we leave in a comment the condition that we
expect is being implicitly checked by the else branch.