Move functions to top of engine pages
This is needed because of a weird quirk with PHPStan, where it would
assume that a function defined after a statement that never returns
(e.g. Page::go or exit) would be unreachable. This is not true, since
PHP parses the entire page before execution, but we make the change
for PHPStan's sake.

Fixes the following (false positive) warning:

> Unreachable statement - code above always terminates.