Refactor Page class (#1355)
Changes to the Page interface
=============================

There is now only one "file" associated with a Page. Previously, this
was achieved with the "url" and "body" array keys. In most cases, the
"url" was either `skeleton.php` or the "body" was empty, so there was
only one real piece of information anyway. In rare cases where we were
using both "url" and "body", this was for directing the "url" page to
forward to the "body" page. However, this can be achieved by passing
the forwarding page in the page data, e.g.

  Page::create('file1', ['forward_to' => 'file2'])

This allows us to remove `skeleton.php` from all Page instantiations.
(In the future, we may convert all engine pages to classes, and use
inheritance to distinguish between processing and display pages, so
that we can better typehint the Page::process return value.)

Require that the "file" be specified at Page construction time. This
allows us to ensure that the appropriate page loads remaining value is
used, and is _always_ present in the instance.

Changes to the Page internals
=============================

Disentangle the Session and Page as much as possible, by limiting the
amount of modifications the Session can make to any Page objects to
just `remainingPageLoads`.

The Session is now responsible for tracking the previous request time,
and we can simply use its `lastAccessed` property (though we did have
to convert it from int to float to recover microtime accuracy for the
page load delay calculation). It simply did not make sense to include
this data in _every_ Page container, especially since there were very
subtle procedures for transfering the request time between intermediate
containers (e.g. going from a processing page to a display page).

Page metadata are no longer stored in the array data, but instead as
properties of the Page object itself. This makes it straightforward to
compute the "Common ID" of the Page, which we now do on the fly. This
is still a bit unsafe, because we can change the Page array data at
any time, but once it is copied into the Session, it should not be
modified any further.