Remove special microtime database handling
There were two motivating factors in the removal:

* We only used the `{get,escape}Microtime` database methods for one
  specific purpose: logging with microsecond precision.

* The implementation, while clever, is somewhat misleading, because
  it's not obvious at first glance that the number stored in the
  database is the actual epoch multiplied by 1e6.

We choose to use a double type to represent this data, for accuracy
of the value in the database and for consistency with other parts of
the code that are dealing with "microtime" values (e.g. the Session
`last_accessed` time).

PHP's `microtime` function doesn't technically have microsecond
precision because a float in PHP typically has only 14 digits of
precision. Thus since we really don't need more than millisecond
precision in the log console, only display up to milliseconds.