Refactor VoteSite class
The VoteSite class is now an enum that only handles the definition of
the specific details of each external voting site. The new VoteLink
class handles the business logic of awarding free turns (communication
with the database, caching, etc.). This can't be part of the VoteSite
enum, since enums cannot have properties.

The main purpose of this change is to narrow the type of VoteSite IDs
from arbitrary ints to a small subset of specific integer-backed
values. Now when we pass a VoteSite to a `match` expression, we can
easily tell if all possible values have been handled.