phpcs.xml: enable parallel processing
Using `--parallel=32` with phpcs reduces the total execution time from
40 seconds to 17 seconds. The parallelization is poor because phpcs is
dominated by a few slow sniffs.

We need to enable the "process control" (pcntl) extension in PHP to
use this feature. Note that pcntl functions should _not_ be used in a
webserver environment, so it is only installed for development builds.