Throw exceptions by default for unhandled match cases
Match expressions that are covered by all the cases should not supply
a `default` case, since it will be unreachable code. This is provable
by static analysis.

For match expressions that are not _provably_ covered by the cases,
we could just rely on the UnhandledMatchError expression, but to show
that we have considered the unhandled cases, we should provide an
explicit `default` case that throws an Exception. Even though the
outcome is the same (an Exception), PHPStan is only satisfied when all
possible cases are covered.

This fixes "Match expression does not handle remaining value" errors.