UniGen: fix infinite loops when populating sectors
To find a valid sector to add locations/warps/planets/ports, we were
looping over all sectors in a galaxy and checking each sector against
a set of conditions.

Since we were not removing invalid sectors from the list as they were
checked, we could run into two problems:

1. Inefficiency, because the same (invalid) sector could be selected
   randomly any number of times before finally selecting a valid one.

2. Infinite loops, because there may be _no_ valid sectors that can
   satisfy the restrictions for selecting a sector.

To fix this, we now remove sectors from the list of eligible sectors
when we determine them to be ineligible. Then when the list of eligible
sectors becomes empty, we throw a UserError.