bootstrap.php: improve exception logging
The new Page class design (see #1355) is not properly displayed with
`var_export` -- it doesn't show any of the private properties of the
class instance, even though the official documentation says it should.
I'm guessing this is an issue specific to ArrayObject.

To fix this, we use `print_r` instead, which does actually display all
properties of the Page class. The downside is that it doesn't display
them _verbatim_ (e.g. it will convert both `null` and `false` to the
empty string, and `true` to "1"), so we may want to switch back to
`var_export` or `var_dump` once Page no longer uses ArrayObject.