bootstrap.php: use <pre> to print exception log
When in debug mode, the exception log message is sent to the browser.
We were using `nl2br` to nicely display it on the screen, but I think
it is better to use `<pre>` tags, because this helps to retain the
original formatting of the message much better. As regular HTML, you
often get unwanted word wrapping and you lose indentation.