phpstan: don't warn about game_delete_processing.php
This Admin Tool is indefinitely disabled, but we keep the code around
just in case we decide we want to be able to delete games again in the
future.

The original implementation was really used to _archive_ games (move
them from the live database to a past games database), but we'd more
likely use it to delete incomplete/test games if we ever re-enable it.