SmrPlayer: add methods has{Good,Evil,Neutral}Alignment
Fixes PHPStan warnings:

government.php:
> Comparison operation ">" between int<-99, max> and -100 is always true.
underground.php:
> Comparison operation "<" between int<min, 99> and 100 is always true.

These arose because we check that the player has the proper alignment
to enter these locations, and then part of the check for neutral
alignment was redundant.

We still want to perform the full neutrality check, just in case we
shuffle the logic of this page around. We can do this by calling into
a separate function that performs the full check without incurring a
warning about "always true".