SmrForce: fix PHPStan warning
> Negated boolean expression is always false.

The `SmrForce::shootPlayers` method checks `SmrForce::exists` at the
top of the method, and then again right before returning. PHPStan was
claiming that the second call would have the same value as the first
call, because it was not aware of any impure methods being called in
between the two calls to `exists`. (An impure method is one that can
change the internal state of the class; PHPStan will automatically
treat a method that returns void as impure, otherwise you have to use
the `@phpstan-impure` annotation.)

As it turns out, an impure method _is_ being called between the two
calls to `exists`, but inside another function (`shootPlayerAsForce`).
This should uncache the expected return value of `exists`, but it does
not currently. See https://github.com/phpstan/phpstan/issues/5727.

To fix this, we need to call the impure methods in the same scope as
the calls to `exists`.