CombatWeapon: remove MaxDamage from damage arrays
The MaxDamage field returned from weapon damage methods was only being
used to track rollover damage in the `takeDamage` methods of combat
entities (SmrForce, SmrPlayer, SmrPort, and SmrPlanet).

However, tracking rollover damage is only needed when we're actually
computing how much of each type of damage is done (shield/CD/armour).
We do not need it to be an intrinsic property of weapons, since that
job is already handled by the `Rollover` boolean.

The design is a bit confusing still. It would be better if there were
a separate "Rollover" damage type, and then weapons would either have
Shield/Armour damage or Rollover damage, but not both. Maybe we'll do
this in a future refactor.