Prevent CombatWeapons from being misused
When a CombatWeapon is used in the wrong context (such as a SmrMine
shooting an SmrPort, or an SmrWeapon being used as a SmrForce), throw
an Exception instead of returning a zero-damage result.

This will help prevent misuse, and will also help guide a refactoring
effort, since classes that share so few methods should not be children
of the same base class.