history_games_detail.php: fix phpstan warning
> Match expression does not handle remaining value: non-empty-string

PHPStan can only narrow the type as the result of using `in_array` if
`true` is passed as the 3rd parameter. This forces `in_array` to do a
strict type comparison, and thus PHPStan has a much more precise idea
about what values the needle can take.

Even with a loose type comparison, PHPStan _could_ narrow the type
better than it does in this case, but it's a hard ask when there's a
simple workaround (and since using strict type comparison is a best
practice anyway).