VoteLink: safer logic for awarding free turns
Refactor the VoteLink class to handle possible race conditions or
abuse in a more robust way. The implementation of such protections
was much more obtuse previously (and particularly hard to explain
to static analysis tools, which assume serial execution and thus
complained that the callback had "always false" conditions).

* Merge the `isClicked` and `setFreeTurnsAwarded` methods into one
  atomic database query. It will both check if the row is in the
  "setClicked state" (i.e. `setClicked` has been called), and will
  update the row to declare that free turns will be awarded without
  any worry about race conditions.

* Add a sanity check to the `setClicked` method to make sure that
  the player can't circumvent the timeout between free turns. We do
  both the check and the update within the scope of a `vote_links`
  table lock to avoid race conditions.

Note that when awarding free turns, we do not need to update the
`vote_links` table from within a sector lock, so long as we always
perform lock-scoped or atomic checks on its state prior to updating.