UniGen: fix phpstan warning
> Variable $fedBeacon might not be defined.

When created linked locations, we need to be more careful about how
we use the `$fedBeacon` variable, in case we ever have a sector with
linked locations that don't include a fed beacon.

Note that we removed the explicit call to add the fed beacon to the
current sector because that's already done in the loop over linked
locations just above.