SmrPort: rework match expression in getGoodClassAtLevel
Fixes PHPStan warning:

> Comparison operation ">=" between int<7, max> and 7 is always true.

We could have set the last branch arm to `default`, since it covers
all remaining possible values of `$level`. However, we really only
want to return a value for a valid level for a port, so we instead
enumerate all possible port levels (and throw an Exception if we have
an invalid port level).