phpstan: ignore unfixable warnings
The default value for the config constant `BUG_REPORT_TO_ADDRESSES` is
an empty array (so that we don't send any bug emails by default). Even
though we have registered this constant as dynamic with PHPStan (i.e.
so that PHPStan understands it is not restricted to a single value),
because an empty array is its own PHPStan type, it still assumes that
it can only ever be an empty array.

See https://github.com/phpstan/phpstan/issues/7520.