Gameplay: make ship maneuverability constant
Previously, maneuverability was a function of your _current_ shields,
armour, and CDs (which would decrease as your ship was damaged). This
doesn't make much sense from a theoretical perspective (you still have
the shield generators and hangar bays on your ship even if they aren't
functional), but it also has practical issues, since its dynamic form
makes it hard for players to account for in tactical decisions.

Just as inertia is a property of matter (thanks Bill), maneuverability
is now a property of ships. It will no longer change until you get a
different ship.