SmrShipType: remove levelNeeded property
At some point, an experimental feature was added to ships where you
had to be at least some minimum level to be able to purchase a given
type of ship. (This was probably an SMR 1.5 change.)

While this is an intriguing idea, we currently don't use it at all,
so there is no reason to set the property. I'll leave the database
column (`ship_type.lvl_needed`) for now, just as a reminder that this
capability exists.

Fixes the PHPStan warning:

> Property SmrShipType::$levelNeeded is never read, only written.