Template: make listjsInclude a public property
Since this property was publicly (and trivially) settable through its
setter method, we can simply remove the method and make the property
public instead.

Normally I don't have a problem with setter methods, but this is a
special case because a) it's only used internally within the class
from the template files and b) it's the only setter method.