SmrSector: change sector link match handling
We would like to be able to use an enum to narrow the type of sector
link directions, but cannot because they are used as array keys.

Restricting their possible values via a special PHPStan docstring type
turns out to be unhelpful, because the docstring would need to
propagate to many more function interfaces, and yet we still wouldn't
actually be able to enforce this restriction anywhere (since we would
ultimately need to assume that some input variable was of this type).

Therefore, we fall back to simply throwing an Exception if an invalid
value is used.

Fixes PHPStan warnings of the type:

> Parameter #1 $dir of method SmrSector::setLinkSector() expects 'Down'|'Left'|'Right'|'Up', string given.