ChessGame: remove method undoMove
This method was only used to restore the original board state after
testing if a move puts the player in check. If we save the original
board state instead, then we don't need a complicated "undo" method.

Note that because the `$board` array contains ChessPiece objects, we
need to deep copy it by cloning the underlying objects.

Many of the functions were passing around the ChessGame properties by
reference _only_ to perform this move/undo logic. By removing this
logic, we no longer need to pass by reference in most cases.