Add Chess\Colour enum
Replace the `ChessGame::PLAYER_*` constants with the Chess\Colour enum.
This narrows the type for many method arguments, which makes the Chess
interface more robust.

Note that the `hasMoved` property uses the colour as an array key, and
enums are not allowed to be array keys. So we use the enum _value_ as
the array key in this case, which is acceptable only because a) this
usage is self-contained and b) the `hasMoved` property needs to be
redesigned anyway.