Make SmrAccount non-abstract
Rename AbstractSmrAccount -> SmrAccount and remove "abstract" marking.
The only use of the abstract base class was to override account newbie
status in Semi Wars games, which has just been offloaded to SmrPlayer.

With this change, the intention going forward is to prevent overriding
behavior of the SmrAccount class for special game types.