RouteGenerator: simplify nested array initialization
Though it is not technically necessary to initialize arrays before
appending to them in PHP, static analysis still encourages us to
initialize non-nested arrays (otherwise, an existing variable with the
same name could accidentally be reused).

This is not the case for nested arrays when the outer array has been
initialized, since it is not possible for the inner array to start out
non-empty in this case.