RouteGenerator: use PHPStan numeric-string typehint
Since array keys can only be strings or ints in PHP, we convert floats
(the route multipliers) to strings so that they can be used as array
keys. As a result, the `numeric-string` type is appropriate here.

One subtlety is that if you use a numeric string as an array key, and
it turns out that the number is an int, then PHP will convert it from
a string to an int! This makes tracking the array key types a bit more
complicated.