composer.json: set config.allow-plugins explicitly
By default, Composer does not run plugins. However, if a package tries
to run a plugin, Composer will output a warning (and even fail in some
cases, e.g. Scrutinizer-CI). We can suppress this warning by including
the package in the `allow-plugins` config.

This fixes the following Composer warning:

> dealerdirect/phpcodesniffer-composer-installer contains a Composer
> plugin which is blocked by your allow-plugins config. You may add it
> to the list if you consider it safe.
> You can run "composer config --no-plugins allow-plugins.dealerdirect
> /phpcodesniffer-composer-installer [true|false]" to enable it (true)
> or disable it explicitly and suppress this exception (false)
> See https://getcomposer.org/allow-plugins

This package is included by `slevomat/coding-standard`. If enabled, it
would do some pre-processing to automatically set the PHP_CodeSniffer
ruleset path to all the ruleset files that it finds in the packages
installed by Composer.

We don't need this feature because we explicitly include rulesets in
our phpcs.xml file. Therefore, we disable plugins for this package.