Use PHPStan template type for array_rand_value
Note: PHPStan doesn't seem to understand the type of function-scoped
static variables, which results in not being able to deduce the
template type for `array_rand_value` in channel_msg_8ball.php. To fix
this, we use a constant instead of a static variable (even though the
static is a nicer solution).