SmrPlayer: fix claimMissionReward error check
There is an unintuitive behavior of the reference assignment operator
where if it references an array element that doesn't exist, the element
is created and set to NULL.

The existing check for strict equivalance to `false` is therefore never
possible (it should have checked for NULL). However, we don't actually
want to be modifying the `missions` array by adding elements in this
way, even if they are NULL. Furthermore, PHPStan does not properly
handle this peculiarity of reference assignment and wrongly reports
that the reference can't be NULL.

The simplest solution is to perform a `hasMission` check _before_ the
reference assignment to avoid all these subtleties.

This fixes PHPStan warning:
> Strict comparison using === between array<string, mixed> and false will always evaluate to false.