shop_hardware_processing.php: add var typehint
PHPStan has no way of knowing the type of an object that we get from
the `$var` (at least until we refactor engine files to take vars as
parameters), so we need to explicitly typehint it.

Note that for local variables, this `@var` docblock typehint overrides
any other information PHPStan has about the variable.