notify_reply.php: simplify recipient name display
Generate the full To/From names in the engine file so that we can
unconditionally print the string in the template.

Also reformatted the displayed name from `<login> a.k.a <player>` to
`<player (Login: <login>)` so that it's easier to figure out who we
are replying to. Made the same change in notify_view.php for
consistency.

Fixes PHPStan warning:

> Call to an undefined method object::getDisplayName().