Handle active player status consistently
Rename existing TIME_BEFORE_INACTIVE constant to TIME_BEFORE_HIDDEN
(this is used for hiding players in Fed who haven't moved in 3 days).

Add a new TIME_BEFORE_INACTIVE constant (10 minutes) that is now used
to consistently handle the "active" status of players. This includes:

* Active/moving players on the CPL
* Number of players playing on the "Play Game" page
* Which players to send Global messages to
* Online/offline status in the Alliance roster

We also replace SmrAccount::isLoggedIn (which used 1hr session timeout)
with SmrAccount::isActive (which uses TIME_BEFORE_INACTIVE), and this
new method is used for determining when a player will start to receive
message notifications.