Session: increase timeout from 1 hour -> 2 days
Players need to log in again after the session timeout is reached,
which means that players now need to log in after 2 days of inactivity
(instead of 1 hour).

There are two caveats that may require revisiting this in the future:

1. Active players who access the server more often than every 2 days
   will only need to log in after a game closure (where all sessions
   are purged). This means that some important post-login processing
   will rarely take place for them, such as: seeing the Announcements
   page, deleting expired `player_has_ticker` rows, deleting expired
   messages, and deleting expired CPL tags.

2. Players using a shared computer that forget to log out will more
   easily leak their session for other players. Such players will need
   to remember to click the "Logout" button. While this was a relevant
   issue in the days of internet cafes, it is very uncommon today.

If these issues are addressed properly, then we might even be able to
make the session timeout much longer (a week? a month?).

Note that this means active players who aren't moving can more easily
stay off the CPL (but will still show up as accessing the server).