current_sector.php: fix refresh all forces message
When there are multiple force stacks in a sector, using "Refresh All"
results in an error due to the `sector_has_forces` query returning more
than one record.

This bug was introduced in 17c0bea3e60, where we removed an important
"LIMIT 1" clause. The fix is to simply add that clause back.