Indent all engine files
In preparation for converting the engine files into Page classes, we
indent every file two levels (since the previous content will now be
scoped inside a class method).

We use the following command to perform the indentation:

    sed -i -r '/^(use |<\?php )/! s/^./\t\t&/' <filename>

Adds two tabs at the start of each non-empty line except for the
opening PHP tag (the first line) and any "use" statement.

Free functions that were not going to be class-scoped were manually
de-indented after the command was run on them.