Convert engine files to Page classes
Page Classes
============

Page classes are constructed directly using the name of the class,
rather than identifying them by the engine filename, e.g.

    new CurrentSector();

instead of

    Page::create('current_sector.php');

Page class names mostly match the previous engine filenames (but with
CamelCase instead of snake_case). In some cases, the class names are a
bit different to improve alphabetic grouping or to clarify what it
actually does (e.g. MessageNotify -> MessageReport). As such, display
Pages no longer map 1:1 to template filenames; in the future we may
either rename the template files to match better, or possibly even
integrate them into the Page classes themselves.

Pages have been put into `Account`, `Player`, and `Admin` namespaces,
and inherit from the expected base classes based on the namespace.
Additional considerations about this arrangement:

* Some pages don't fit cleanly into these categories. For example, they
  will include player-related content if the session has a game;
  otherwise they are account-related only (see, e.g. `Preferences`).
  It's not clear how to handle these cases, but we default to using the
  `Account` namespace since it has the most generic interface.

* Some Pages don't consume all the interface arguments inherited from
  their base class (e.g. an `PlayerPage` may not actually need access
  to the `$player` object). This may change with further refactoring
  (e.g. if we pass the arguments along to the Template).

* Sub-namespaces have been created to help organize the massive list
  of pages into logical units (e.g. Planet, Bank, Bar, etc.).

Page Variables
==============

Page variables are now properties of the associated class, and are
accessed directly in the class scope rather than through the session
"var".

We have attempted to make most Page properties `readonly`, since it is
important for Pages to be unchanged once we create an SN for them
(since changing it afterwards would unexpectedly change any other links
that shared that SN when they were originally identical objects).
However, in some cases we could not make every property `readonly` in a
a sensible way, and further refactoring is needed to achieve this.

Since we want to make the Page properties readonly whenever possible,
we sometimes run into the issue that we want to clone a Page, but
modify one of the readonly properties. This is not possible in PHP at
this time, but see https://github.com/spatie/php-cloneable.

For the sake of simplicity, many conditionally set Page properties are
now _always_ assigned to the Template. Since they are now `null` rather
than undefined, it will be treated as if it weren't assigned by the
Template (due to `isset(null) === false`). This does change the _keys_
of the Template data array, but we only ever care about their _values_.

Additional Changes
==================

In many cases, the conversion was straightforward and did not change
any functionality. However, many inconsistencies or other minor bugs
were found and fixed as part of the conversion process. Ideally, any
bugfixes would be made in a separate commit from such a large refactor,
but they were often fundamentally a consequence of the confusing and
implicit nature of the "container" design. Below is a sample of the
functional/organizational changes that were made:

* SectorsFile is converted to a regular library class, so that it can
  be reused by both the admin and player pages that create Sectors
  files.

* AlbumDeleteProcessor returns to AlbumEdit instead of CS/Game Play.

* Fix HallOfFamePersonal when the session player has not joined the
  game being viewed.

* HistoryGames pages inherit from an abstract class `HistoryPage` that
  handles common functionality between all HistoryGames pages, such as
  switching between live/history databases, and adding the Menu with
  the current page bolded without a hard-coded order number.

* More robust way of detecting if we're processing a page related to
  unbanning the session account. We do this by checking the class of
  the page we're forwarding to (e.g. `InvalidEmailProcessor` and
  `ReopenAccountProcessor`).

* Break up LoginCheckValidatedProcessor into multiple classes to make
  the login processing flow more straightforward.

* LoginProcessor goes straight to GamePlay, since that will forward to
  CurrentSector if the session is already in a game.

* Add NewsReadAdvancedProcessor so that we're not making redundant
  database queries every time we refresh the NewsReadAdvanced page.

* Add missing confirmation message when changing color preferences.

* Move credit transfer processing out of PreferencesProcessor and into
  its own separate PreferencesTransferProcessor page.

* Add separate NpcManageAddAccountProcessor. Having separate Processors
  is another way to avoid depending on hard-coded submit/input values
  to trigger which type of processing happens.

* Move some logic out of UniGen\SaveProcessor for processing requests
  from completely different pages into their own Processor pages.
  Ideally, this would be broken up further, but there were some shared
  functions that need to be refactored first.

* Fix "Refresh All" message disappearing after all forces refreshed.

* Add sanity checks to Draft-only Pages that the game mode is actually
  Draft.