Add infrastructure for Page classes
The Page class will no longer provide a generic container for storing
variables associated with engine pages. Instead, each engine page will
be converted into a class (deriving from Page) that specifies in its
constructor interface exactly which variables it expects to store.

The main reason for this change is that generic containers make it
entirely impossible to use static analysis to determine if we have
provided the proper data for any Page we're linking/forwarding to.

Another problem with the old design was simply referring to the page.
Since it just used a filename, there was no way to validate that we
were constructing a page that actually existed without running it.

Old engine pages also had a lot of non-injectible boilerplate at the
top of each file to load standard variables (template, account, player,
etc.). The new Page classes will inherit from one of the following
abstract specializations:

* PlayerPage
* AccountPage
* PlayerPageProcessor
* AccountPageProcessor

Each provides a different interface for building the page, tailored
to the context in which it is used. Processor pages formalize the
distinction between display pages (which return) and processing pages
(which do not; they always forward to another Page).

User request data obviously cannot be known at Page construction time,
but it still needs to be stored so that it can be reused on AJAX auto
page refreshes. There are two ways to achieve this:

1. Link to a Processor page, which will then construct a display Page,
   with all the request data stored as class properties.

2. Store it in the new `Session::requestData` property, which is very
   similar to how the old Page class worked (just a generic container
   to hold unknown variables). This is sometimes helpful when it is too
   much boilerplate to create a whole Processor page just to store a
   simple Request variable. We may want to phase this out.

Page reusability is now specified locally for each Page class, rather
than storing it in the `Page::ALWAYS_AVAILABLE_PAGES` array. This
improves encapsulation, and prevents this property from getting out of
sync with the intended design of any particular Page. We implement this
with the `ReusableTrait` class; if not used we default to the Page not
being reusable.

There are a few properties declared in the base Page that we are not
handling in an ideal way. These are `$skipRedirect`, `$allowAjax`, and
`$underAttack`. If we ever want Page classes to be `readonly`, we will
need to find a different way to handle these concepts. Note that the
latter two replace the prior `AJAX` and `UnderAttack` page variables
respectively.

Another deficiency is that we would like `$file` to only be declared
for display pages, but it currently needs to exist for all Page classes
to satisfy PHPStan.