Database: more robust close/reconnect procedures
Replace the `close` and `reconnect` methods with a static method
`resetInstance`. This restores the "on demand" aspect of the Database
class in cases where we have to close the database connection (i.e.
for persistant services such as the Discord/IRC clients).

Previously, reconnection was done manually to avoid creating overhead
in the `getInstance` method, which would slow down the traditional
usage by the game engine (where we have one connection per page). Now,
instead of just closing the connection, we purge the entry from the
DI container as well, so that the next time `getInstance` is called,
we simply build a new database connection from scratch.

As a result, we avoid unnecessary reconnections in many cases.

Note that we are closing the database connection explicitly (rather
than just unsetting the reference in the DI container and letting it
close on garbage collection) because we can't be sure that some other
class (e.g. SmrAccount, SmrAlliance, etc.) hasn't saved a reference to
the Database.