Add Blackjack Table class
This class simplifies the engine processing by holding the deck,
player hand, and dealer hand. It provides useful methods for actions
taken during a game ("hitting"), and introspection of the state of
the table (e.g. is the game over? did the player win?).

The Result enum provides a better way to assert the result of the
game instead of relying on string comparison (yes/no/tie/bj).