bar_gambling_processing.php: convert to template
Previously, this processing page was creating the entire contents of
the display page, and then passing that in the container to that page.
This was detrimental for several reasons:

1. It prevented the use of the template mechanism.

2. The links displayed on the page could not be updated (since the
   processing page could not be run again), and we needed a hack to
   prevent AJAX from running while playing Blackjack.

The new design standardizes this page: only the processing is done in
the processing page, and all the display logic is done in the display
page (along with a template). The AJAX hack in `do_voodoo` is also
removed, and AJAX works properly now while playing Blackjack.