Move all engine files into the Smr\Pages directory
This commit is separate only for the purposes of improving git rename
detection. These files will be significantly edited, and doing that at
the same time as a rename can disassociate it with the original file.

Since most of the changes will be indentation changes, the diffs are
still very meaningful.

The `SectorsFile` class is a special case, because we're moving it to
the `Smr` namespace instead of `Smr\Pages`.