Fix user error handling
The `create_error` function is now just a wrapper around the UserError
exception so that all user errors are fundamentally exception-based.
This will allow us to execute engine code and handle errors in custom
ways depending on context (e.g. in tests or CLI scripts).

As such, in `handleException`, rather than calling `create_error` when
handling a UserError exception, the `handleUserError` function will be
called, which contains the previous logic of `create_error`.

Eventually, we may remove the `create_error` wrapper in favor of just
throwing UserError directly. But for now, the wrapper simplifies this
transition.

The error.php engine page now _only_ displays the error page, rather
than being overloaded to display the error on the Current Sector page
if appropriate. That logic has been moved into `handleUserError` so
that error.php is now a pure display page.

This fixes several broken features:

1. If an error occurs during AJAX auto-refresh, previously the page
   would often stop refreshing (which could be very confusing). Now,
   it will always forward to an error page. This was nominally added
   in a0d24c0e990, but didn't always work properly due to the Template
   not be being reset during error handling. Then 6e7995cab claimed to
   fix this, but only fixed part of the problem.

2. The NPC script was failing in an unhelpful way ("Cannot modify
   header information") when it would hit `create_error` during the
   processing of a container. This is because you can't use `header()`
   in a CLI script. Now, hitting `create_error` will throw a UserError
   exception, which the NPC script can catch and handle properly.

Additional notes:

* We can't require a sector lock when displaying an AJAX error on the
  Current Sector page, because AJAX updates don't have sector locks
  (which should be fine, because auto-refresh shouldn't be modifying
  data anyway).

* We can remove the similar AJAX redirection in planet.inc.php when a
  player is kicked from a planet, and instead just throw an error to
  reuse the AJAX redirection code in `handleUserError`. Note that it
  then calls it an "error" instead of a "warning", but I think that's
  a fair price for code deduplication.